# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cybersecurity Dork Library API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dorks:
    get:
      operationId: listDorks
      tags: [dorks]
      summary: List dorks with filters and pagination
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: category
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: platform
          in: query
          schema:
            type: string
        - name: operator
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
        - name: intentType
          in: query
          schema:
            type: string
        - name: sourceType
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [newest, popular, copied, difficulty]
            default: newest
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of dorks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DorkListResponse"
    post:
      operationId: createDork
      tags: [dorks]
      summary: Create a new dork
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDorkBody"
      responses:
        "201":
          description: Created dork
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dork"
  /dorks/{id}:
    get:
      operationId: getDork
      tags: [dorks]
      summary: Get a dork by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dork detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DorkDetail"
    put:
      operationId: updateDork
      tags: [dorks]
      summary: Update a dork
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDorkBody"
      responses:
        "200":
          description: Updated dork
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dork"
    delete:
      operationId: deleteDork
      tags: [dorks]
      summary: Delete a dork
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /dorks/trending:
    get:
      operationId: getTrendingDorks
      tags: [dorks]
      summary: Get trending dorks
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Trending dorks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dork"
  /dorks/recent:
    get:
      operationId: getRecentDorks
      tags: [dorks]
      summary: Get recently discovered dorks
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent dorks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dork"
  /dorks/{id}/related:
    get:
      operationId: getRelatedDorks
      tags: [dorks]
      summary: Get related dorks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Related dorks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dork"
  /dorks/{id}/versions:
    get:
      operationId: getDorkVersions
      tags: [dorks]
      summary: Get version history for a dork
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dork versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DorkVersion"
  /dorks/{id}/copy:
    post:
      operationId: recordDorkCopy
      tags: [dorks]
      summary: Record a dork copy event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Copy recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    Dork:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        queryTemplate:
          type: string
        optimizedQuery:
          type: string
        description:
          type: string
        usageContext:
          type: string
        intentType:
          type: string
        difficulty:
          type: string
        sourceType:
          type: string
        status:
          type: string
        viewsCount:
          type: integer
        copyCount:
          type: integer
        primaryCategoryId:
          type: integer
        primaryCategory:
          $ref: "#/components/schemas/Category"
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        operators:
          type: array
          items:
            $ref: "#/components/schemas/Operator"
        platforms:
          type: array
          items:
            $ref: "#/components/schemas/Platform"
        categories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, queryTemplate, status, createdAt]
    DorkDetail:
      allOf:
        - $ref: "#/components/schemas/Dork"
        - type: object
          properties:
            versions:
              type: array
              items:
                $ref: "#/components/schemas/DorkVersion"
            relatedDorks:
              type: array
              items:
                $ref: "#/components/schemas/Dork"
    DorkListResponse:
      type: object
      properties:
        dorks:
          type: array
          items:
            $ref: "#/components/schemas/Dork"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [dorks, total, page, limit, totalPages]
    CreateDorkBody:
      type: object
      properties:
        title:
          type: string
        queryTemplate:
          type: string
        optimizedQuery:
          type: string
        description:
          type: string
        usageContext:
          type: string
        intentType:
          type: string
        difficulty:
          type: string
        sourceType:
          type: string
        status:
          type: string
        primaryCategoryId:
          type: integer
        categoryIds:
          type: array
          items:
            type: integer
        tagIds:
          type: array
          items:
            type: integer
        operatorIds:
          type: array
          items:
            type: integer
        platformIds:
          type: array
          items:
            type: integer
      required: [title, queryTemplate]
    DorkVersion:
      type: object
      properties:
        id:
          type: integer
        dorkId:
          type: integer
        queryTemplate:
          type: string
        optimizedQuery:
          type: string
        description:
          type: string
        versionLabel:
          type: string
        changeReason:
          type: string
        changedAt:
          type: string
      required: [id, dorkId, changedAt]
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        icon:
          type: string
        color:
          type: string
        depthLevel:
          type: integer
        displayOrder:
          type: integer
        isActive:
          type: boolean
        parentCategoryId:
          type: integer
        dorkCount:
          type: integer
        subcategories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        createdAt:
          type: string
      required: [id, name, slug, depthLevel]
...
    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        tagType:
          type: string
        description:
          type: string
        dorkCount:
          type: integer
      required: [id, name, slug]
...
    Operator:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        syntax:
          type: string
        description:
          type: string
        exampleUsage:
          type: string
        dorkCount:
          type: integer
      required: [id, name, syntax]
...
    Platform:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        platformType:
          type: string
        description:
          type: string
        dorkCount:
          type: integer
      required: [id, name, slug]
...